Fixing teleport orientations
Fixed the pitch/rotation confuzzling with Location and teleportTo
Removed setDimension from entity as it does not work
without sending an update packet to a player. Also it doesn't change an Entities position.
This should not have been exposed.
Use teleportTo(Location) or teleportTo([...], World) instead.
This will automatically switch worlds if required and do all the packet sending work.